requirements.txt
flask>=2.2.5
flask-sqlalchemy>=3.0.5
transformers>=4.30.0
torch>=2.0.0
python-dotenv==1.0.0
pandas==1.5.3
numpy==1.24.3
plotly==5.15.0
openai>=1.0.0
Pillow>=10.0.0
moviepy>=1.0.3
imageio>=2.31.1
imageio-ffmpeg>=0.4.8 